requirements.txt
Django==4.0.2
argparse==1.2.1
factory-boy==2.4.1
pytz==2014.10
python-coveralls==2.5.0
coverage==3.7.1
psycopg2==2.9.3
djangorestframework==3.13.1
drf-spectacular==0.21.2
drf-spectacular[sidecar]
djangorestframework_simplejwt==5.0.0
django-redis==5.2.0


